requirements.txt
# Truck Scale Weighing System Dependencies

# Core GUI and Database
tkinter  # Built-in with Python
sqlite3  # Built-in with Python

# Serial Communication
pyserial>=3.5

# Camera Support
opencv-python>=4.5.0
Pillow>=8.0.0

# PDF Generation
reportlab>=3.6.0

# Excel Export (for enhanced CSV export with auto-fit)
openpyxl>=3.0.0

# Optional: Advanced logging
colorlog>=6.0.0

# Optional: System tray support (if needed)
pystray>=0.19.0

# Optional: Better error handling
traceback2  # For older Python versions compatibility
